Refactor capture mechanics to be in a seperate component
